+++ project/libs/shared/core/src/base/entity.ts
