+++ project/libs/shared/core/src/constants/common.ts
