+++ project/libs/shared/core/src/constants/user.ts
