+++ project/libs/shared/core/src/index.ts
@@ -1,9 +1,9 @@
-export { Entity } from './lib/base/entity';
+export { Entity } from './base/entity';
 
-export type { User } from './lib/types/user.interface';
-export type { AuthUser } from './lib/types/auth-user.interface';
-export type { StorableEntity } from './lib/types/storable-entity.interface';
-export type { EntityFactory } from './lib/types/entity-factory.interface';
+export type { User } from './types/user.interface';
+export type { AuthUser } from './types/auth-user.interface';
+export type { StorableEntity } from './types/storable-entity.interface';
+export type { EntityFactory } from './types/entity-factory.interface';
 
-export { SALT_ROUNDS, AUTH_USER_EXISTS } from './lib/constants/user';
-export { ENTITY_NOT_FOUND } from './lib/constants/common';
+export { SALT_ROUNDS, AUTH_USER_EXISTS } from './constants/user';
+export { ENTITY_NOT_FOUND } from './constants/common';
