+++ project/libs/shared/core/src/types/auth-user.interface.ts
