+++ project/libs/shared/core/src/types/entity-factory.interface.ts
