+++ project/libs/shared/core/src/types/storable-entity.interface.ts
