+++ project/libs/shared/helpers/src/common.ts
