+++ project/libs/shared/helpers/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/common';
+export * from './common';
